requirements.txt
pysolarmanv5>=2.4.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
requests>=2.31.0
pycryptodome>=3.19.0
